Create README.md
This will be a test app.  My goal is to generate two rows of red dots representing sales on silver background.
Then press a button and add rows of Green dots along side the red dots.  The green dots come from a second JSON file that I want to add with D3 after the SVG is already drawn.